requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
pymodbus==3.7.2
pyserial==3.5
PyYAML==6.0.2
yarl==1.9.4
